# specs/002-fullstack-web-auth/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Todo Application API
  description: |
    RESTful API for Phase II Todo Application with JWT authentication.

    **Authentication**: All protected endpoints require a valid JWT access token in the Authorization header.

    **User Isolation**: All task operations are filtered by the authenticated user's ID.
  version: 1.0.0
  contact:
    name: Todo App Support
    email: support@todoapp.example.com
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.todoapp.example.com
    description: Production server (Phase V)
...
security:
  - BearerAuth: []
paths:
...
  /api/auth/signup:
    post:
      summary: User registration
      description: Create a new user account with email, name, and password
      tags:
        - auth
      security: []  # Public endpoint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Invalid input (email format, password validation, etc.)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                weakPassword:
                  value:
                    detail: "Password must be at least 8 characters"
                emailExists:
                  value:
                    detail: "Email already registered"
  /api/auth/login:
    post:
      summary: User login
      description: Authenticate user with email and password, returns access and refresh tokens
      tags:
        - auth
      security: []  # Public endpoint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                detail: "Invalid credentials"
  /api/auth/refresh:
    post:
      summary: Refresh access token
      description: Use refresh token to obtain a new access token
      tags:
        - auth
      security: []  # Uses refresh token, not access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
                  example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
              required:
                - refresh_token
      responses:
        '200':
          description: New access token issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                  token_type:
                    type: string
                    example: "bearer"
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /api/auth/login (15 minute expiration)
  schemas:
    UserCreate:
      type: object
      required:
        - email
        - name
        - password
      properties:
        email:
          type: string
          format: email
          example: "alice@example.com"
          description: User email address (must be unique)
        name:
          type: string
          maxLength: 255
          example: "Alice Smith"
          description: User display name
        password:
          type: string
          minLength: 8
          example: "SecurePass123"
          description: Password (min 8 chars, 1 uppercase, 1 lowercase, 1 number)
    UserLogin:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "alice@example.com"
        password:
          type: string
          example: "SecurePass123"
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        email:
          type: string
          format: email
          example: "alice@example.com"
        name:
          type: string
          example: "Alice Smith"
        created_at:
          type: string
          format: date-time
          example: "2025-12-30T10:00:00Z"
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          description: JWT access token (15 minute expiration)
        refresh_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          description: JWT refresh token (7 day expiration)
        token_type:
          type: string
          example: "bearer"
        expires_in:
          type: integer
          example: 900
          description: Access token expiration in seconds (900 = 15 minutes)
...
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
          example: "Invalid credentials"
          description: Error message